+++ ts-tests/containers/MetaDContainer.ts
@@ -1,3 +1,4 @@
+import { ethers } from 'ethers';
 import Web3 from 'web3';
 import { JsonRpcResponse } from 'web3-core-helpers';
 import {
@@ -6,21 +7,7 @@ import {
   StartedNetwork,
   StartedTestContainer
 } from 'testcontainers';
-import { ethers } from 'ethers';
-import { Keyring } from '@polkadot/api';
-
-const keyringEth = new Keyring({ type: 'ethereum' });
-
-// constant
-export const GENESIS_ACCOUNT = '0x6be02d1d3665660d22ff9624b7be0551ee1ac91b';
-export const GENESIS_ACCOUNT_BALANCE =
-  '340282366920938463463374607431768210955';
-export const GENESIS_ACCOUNT_PRIVATE_KEY =
-  '0x99B3C12287537E38C90A9219D4CB074A89A16E9CDB20BF85728EBD97C343E342';
-export const genesis = keyringEth.addFromUri(GENESIS_ACCOUNT_PRIVATE_KEY);
-
-export const CHAIN_ID = 988;
-export const EXISTENTIAL_DEPOSIT = 500;
+import { CHAIN_ID } from '../utils/constant';
 
 type MetaDNetwork = 'mainnet' | 'testnet';
